adding crud on messages and defining reflatioships +adding comment section

+++ base/views.py
@@ -1,11 +1,12 @@
 from django.shortcuts import redirect, render
 from django.http import HttpResponse
-from .models import Room, Topic,User
+from .models import Room, Topic,User,Message
 from django.db.models import Q
 from .forms import RoomForm
 from django.contrib import messages
 from django.contrib.auth import authenticate , login , logout 
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 #create function and methods
 """rooms=[
     {'id':1, 'name':'room1'},
@@ -14,9 +15,12 @@ from django.contrib.auth.decorators import login_required
 ]"""
 
 
-def loginPage(request):    
+def loginPage(request):  
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')  
     if request.method=='POST':
-        username=request.POST.get('username')
+        username=request.POST.get('username').lower()
         password=request.POST.get('password')
         try:
             user=User.objects.get(username=user)
@@ -28,11 +32,31 @@ def loginPage(request):
             return redirect('home')
         else:
             messages.error(request,'Username OR password does not exist')
-    context={}
+    context={'page':page}
     return render(request, 'base/login_register.html',context)
+
+
 def logoutUser(request):
     logout(request)
     return redirect('home')  
+
+
+def registerPage(request):
+    form=UserCreationForm()
+    if request.method=='POST':
+        form=UserCreationForm(request.POST)
+        if form.is_valid():
+            user =form.save(commit=False)
+            user.username=user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error occurrred during registration')
+    context={'form':form}
+    return render (request,'base/login_register.html', context)
+
+
 
 def home(request):
     q=request.GET.get('q') if request.GET.get('q')!=None else '' 
@@ -48,8 +72,21 @@ def home(request):
 #room functions
 def room(request, pk):
     room=Room.objects.get(id=pk)
-    context ={'room':room}
+    room_messages=room.message_set.all().order_by('-created')
+    participants=room.participants.all()
+    if request.method=='POST':
+        message=Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room',pk=room.id)
+    context ={'room':room , 'room_messages':room_messages, 'participants':participants}
     return render(request, 'base/room.html',context)
+
+
+
 @login_required(login_url='login')
 def createRoom(request):
     form=RoomForm()
@@ -61,22 +98,39 @@ def createRoom(request):
     context={'form':form}
     return render(request, 'base/room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)#prefill the form with existing data 
+   
     if request.method=='POST':
        form=RoomForm(request.POST,instance=room)
+       if request.user!=room.user:
+        return HttpResponse('you are not allowed here') 
        if form.is_valid():
            form.save()
            return redirect('home')
     context={'form':form}
     return render(request, 'base/room_form.html',context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user!=room.user:
+        return HttpResponse('you are not allowed here')
     if request.method=='POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'object':room})
 
-#topic functions
+@login_required(login_url='login')
+def deleteMessage(request,pk):
+    message=Message.objects.get(id=pk)
+    room=message.room
+    if request.user!=message.user:
+        return HttpResponse('you are not allowed here')
+    if request.method=='POST':
+        message.delete()
+        return redirect('room',room.id)
+    return render(request, 'base/delete.html', {'object':message})
